add docker-compose.yml, exceptions, application.yml and logic

+++ src/main/java/com/company/generatecheck/dto/ItemDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class ItemDto {
 
-    private Short qty;
+    private Integer qty;
     private ProductDto product;
     private Integer total;
 }
